requirements.txt
aiohttp==3.11.10
redis==6.0.0
async-timeout==5.0.1
beautifulsoup4==4.12.3
fastapi==0.115.12
fake-useragent==2.0.3
feedparser==6.0.11
google-search-results==2.4.2
markdownify==0.14.1
nest_asyncio==1.6.0
openai==1.66.3
pandas==2.2.3
psutil==7.0.0
pydantic==2.11.4
python-dotenv==1.1.0
pyppeteer==2.0.0
streamlit==1.45.0
tqdm==4.67.0
tiktoken==0.8.0
uvicorn==0.34.2 